% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMOLE.R
\name{dMOLE}
\alias{dMOLE}
\alias{pMOLE}
\alias{qMOLE}
\alias{rMOLE}
\alias{hMOLE}
\title{The Marshall-Olkin logistic-exponential distribution}
\usage{
dMOLE(x, mu, sigma, nu, log = FALSE)

pMOLE(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qMOLE(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rMOLE(n, mu, sigma, nu, tau)

hMOLE(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMOLE} gives the density, \code{pMOLE} gives the distribution
function, \code{qMOLE} gives the quantile function, \code{rMOLE}
generates random deviates and \code{hMOLE} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the Marshall-Olkin logistic-exponential distribution
with parameters \code{mu}, \code{sigma}, and \code{nu} .
}
\details{
The Marshall-Olkin logistic-exponential Distribution with parameters \code{mu},
\code{sigma}, and \code{nu} has density given by

\eqn{f(x)=\ [\mu \sigma \nu \exp^{\nu x}[\exp^{\nu x}-1]^{-\mu-1}/ [1 + \sigma [\exp^{\nu x}-1]^-\mu]^2}

for x > 0
}
\examples{

## The probability density function
par(mfrow=c(1, 2))
 curve(dMOLE(x, mu=0.6, sigma=5.5, nu=1), from=0, to=8,
 ylim=c(0, 0.3), col="red", las=1, ylab="f(x)")

 curve(dMOLE(x, mu=3, sigma=15, nu=1.2), from=0, to=3,
 ylim=c(0, 1.5), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pMOLE(x, mu=0.6, sigma=5.5, nu=1),
      from=0, to=8,  col="red", las=1, ylab="F(x)")
curve(pMOLE(x, mu=0.6, sigma=5.5, nu=1, lower.tail=FALSE),
      from=0, to=8, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qMOLE(p, mu=0.6, sigma=5.5, nu=1), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pMOLE(x, mu=0.6, sigma=5.5, nu=1), from=0, add=TRUE, col="red")

## The random function
hist(rMOLE(n=10000, mu=0.6, sigma=5.5, nu=1), freq=FALSE,
     xlab="x", las=1, main="")
curve(dMOLE(x,  mu=0.6, sigma=5.5, nu=1), from=0, add=TRUE, col="red")

## The Hazard function
par(mfrow=c(1,1))
curve(hMOLE(x,  mu=0.6, sigma=5.5, nu=1), from=0, to=8,
      col="red", ylab="Hazard function", las=1)

}
