% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafvp.R
\name{grafvp}
\alias{grafvp}
\title{Gráfico de valores p en t.test}
\usage{
grafvp(x)
}
\arguments{
\item{x}{vector de valores p con la función \code{\link{valp}}}
}
\description{
Gráfico de valores p obtenido con valp.
}
\examples{
grafvp(valp()) # 100 comparaciones por defecto
grafvp(valp(cp = 10)) # modificando el número de comparaciones

}
