% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valp.R
\name{plot.valp}
\alias{plot.valp}
\title{Gráfico de valores p en t.test}
\usage{
\method{plot}{valp}(x)
}
\arguments{
\item{x}{vector de valores p con la función \code{\link{valp}}}
}
\description{
Gráfico de valores p obtenido con valp.
}
