% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valp.R
\name{valp}
\alias{valp}
\title{Valor p en múltiples t.test}
\usage{
valp(n = 1000, gp = 20, cp = 100, nc = 0.95)
}
\arguments{
\item{n}{tamaño poblacional (por defecto igual a 1000)}

\item{gp}{tamaño muestral (por defecto igual a 20)}

\item{cp}{número de comparaciones (por defecto igual a 100)}

\item{nc}{nivel de confianza (por defecto igual a 0.95)}
}
\value{
un vector con los valores p obtenidos para las comparaciones
establecidas.
}
\description{
Esta función genera una población de números aleatorios con
distribución normal y con base en ella se realizan múltiples
muestreos; luego se emplea el t.test para obtener el número de veces que
el contraste de hipótesis genera conclusiones erradas.
}
\details{
Detalles...
}
